#2 - refactoring; getting assessment rubric
Renaming some of the variables because there are two kinds of rubric involved.  One rubric is for the assignment and seems to be used by the instructors for grading.  The other rubric is the assessment rubric used by reviewers (assessors) for writing a review comments during a peer review (assessment).