#2 - assessment rubric ID found
The assessment rubric ID is found in a couple places in each assessment.  The same rubric is referenced many times for each assignment, so it was decided to try the logic shortcut of using the rubric from the first assessment and assume it's the same for all the other assessments.