#3 - get and store rubric criteria
This was a little trickier because Canvas' data uses ID *_strings_* for the RubricCriterion objects, with the format `_𝒏𝒏𝒏…`.  After a while, I decided that it would be worthwhile to take the risk and assume that the ID with the underscore stripped off would be unique among all criteria, not only those within each rubric.