#3 - use specific application name with migrations
As pointed out by @zqian while reviewing #5, migrations weren't being created.  In the future, after this PR is merged, migrations will become part of the repo, as is standard practice, but for now, they're being generated as needed.  This had been working until recently.  In fact, it's still working with my development copy, but when I checked out a fresh copy, I was able to reproduce the problem.  Perhaps my upgrading the project to Python 3.11 or some other Django upgrades changed how migrations are made.